ESWE-791_prisoner_view_cms Minor refactor, tidy
