+++ server/routes/prisonerListMatchJobs/index.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express'
-import PrisonerListMatchedJobsController from './prisonerListMatchedJobsController'
+import PrisonerListMatchJobsController from './prisonerListMatchJobsController'
 import type { Services } from '../../services'
-import getPrisonerListMatchedJobsResolver from '../../middleware/resolvers/getPrisonerListMatchedJobsResolver'
+import getPrisonerListMatchedJobsResolver from '../../middleware/resolvers/getPrisonerListMatchJobsResolver'
 import handleSortMiddleware from '../../middleware/handleSortMiddleware'
 
 export default (router: Router, services: Services) => {
-  const controller = new PrisonerListMatchedJobsController(services.prisonerSearchService, services.paginationService)
+  const controller = new PrisonerListMatchJobsController(services.prisonerSearchService, services.paginationService)
   router.get('/cms/prisoners', [getPrisonerListMatchedJobsResolver(services.prisonerSearchService)], controller.get)
 
   router.post('/cms/prisoners', [handleSortMiddleware('sortAction', 'releaseDate')], controller.post)
